Initial commit
This is a stand-alone plant model. The model is compiled as library for use by the joint ecological model, or as a standalone. Either of those may become CAD-integrated.
Note that the code is not for sharing or distribution, despite GPL statements in individual files! see readme.md.
Model is not yet ready for use, see section important caveats in readme.md,