update documentation
large parts of the readme are updated, so that this documentation only concerns the plant model (with occasional refs to the joint model). Concerns issue #1 but does not solve it fully. To do: Figures from TEAMS doc to insert; finish description of germination and reproduction; write chapter 8; include text from major #issues; write USER_GUIDE.md